gateway/requirements.txt
psycopg2-binary
pyserial
XBee
